fire/requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
bcrypt==4.2.0
billiard==4.2.0
boto3==1.35.21
botocore==1.35.21
ccxt==4.4.3
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
dependency-injector==4.42.0
fastapi==0.114.2
flower==2.0.1
fonttools==4.54.1
frozenlist==1.4.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.10.0
idna==3.10
iniconfig==2.0.0
jmespath==1.0.1
jwt==1.3.1
kiwisolver==1.4.7
kombu==5.4.1
magic-filter==1.0.12
matplotlib==3.9.2
multidict==6.1.0
numpy==2.0.2
orjson==3.10.7
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==11.0.0
pluggy==1.5.0
prometheus-client==0.20.0
prompt-toolkit==3.0.47
pybit==5.8.0
pycares==4.4.0
pycparser==2.22
pycryptodome==3.21.0
pycurl==7.45.3
pydantic==2.9.1
pydantic-core==2.23.3
pydantic-settings==2.5.2
pyjwt==2.9.0
pyparsing==3.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
redis==5.0.8
requests==2.32.3
s3transfer==0.10.2
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.34
starlette==0.38.5
tornado==6.4.1
typing-extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.11.1
